refactor TUI modules in `src/tui/`
